requirements.txt
pandas==2.2.3
SQLAlchemy==2.0.37
pyodbc==5.2.0
python-dotenv==1.0.1
matplotlib==3.10.0
seaborn==0.13.2
pytest==8.3.4
